+++ src/io/ProjectFile.js
@@ -19,6 +19,7 @@
  */
 
 import { CAPTURE_INDIRECT_BASE, migrateCaptureBase, migrateSdfParams,
+         migrateScene3dParams,
          PARAM_SCHEMA, migrateBlendPercent } from '../controls/ParameterSystem.js';
 
 const FORMAT_VERSION = 3;
@@ -229,6 +230,7 @@ export class ProjectFile {
     if (data.params) {
       migrateCaptureBase(data.params, data._sourceCount);
       migrateSdfParams(data.params);
+      migrateScene3dParams(data.params);
       migrateBlendPercent(data.params, null, data._schema);
       this.ps.restoreState(data.params);
     }